+++ src/main/java/com/guess/foodback/authOrder/OrderRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
 
-    Order findFirstByUser_IdAndStatus(Long userId, Status status);
+    Order findFirstByUser_IdAndStatus(Integer userId, Status status);
 
 }
